Move azureauth to upper package
