+++ pkg/tsdb/prometheus/azureauth/azure.go
