+++ pkg/tsdb/prometheus/azureauth/azure_test.go
